Add 'Contact List' tab to user details
