+++ app/features/contact/index.ts
@@ -1,1 +1,2 @@
 export * from './components/contact-form';
+export * from './components/contact-list';
