+++ app/routes/email-activity/index.tsx
@@ -1,7 +1,6 @@
 import type { Route } from './+types/index';
 import { EmailList } from '@/features/email';
 import { emailListQuery } from '@/resources/request/client';
-import { routes } from '@/utils/config/routes.config';
 import { metaObject } from '@/utils/helpers';
 import { Trans } from '@lingui/react/macro';
 
@@ -14,11 +13,5 @@ export const handle = {
 };
 
 export default function Page() {
-  return (
-    <EmailList
-      queryKeyPrefix="emails"
-      fetchFn={() => emailListQuery('milo-system', {})}
-      detailPath={(namespace, emailName) => routes.emailActivity.detail(namespace, emailName)}
-    />
-  );
+  return <EmailList queryKeyPrefix="emails" fetchFn={() => emailListQuery('milo-system', {})} />;
 }
