--- app/routes/user/detail/email-activity/layout.tsx
@@ -1,10 +1,0 @@
-import { Trans } from '@lingui/react/macro';
-import { Outlet } from 'react-router';
-
-export const handle = {
-  breadcrumb: () => <Trans>Email Activity</Trans>,
-};
-
-export default function Layout() {
-  return <Outlet />;
-}
